refactor: 예약 컬렉션 인터페이스화
