+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.JdbcReservations;
 import roomescape.domain.Reservation;
+import roomescape.domain.Reservations;
 
 import java.net.URI;
 import java.util.List;
@@ -18,7 +19,7 @@ import java.util.List;
 @RequestMapping("/reservations")
 public class ReservationController {
 
-    private final JdbcReservations reservations;
+    private final Reservations reservations;
 
     public ReservationController(JdbcReservations reservations) {
         this.reservations = reservations;
