+++ src/main/java/roomescape/domain/JdbcReservations.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Component
-public class JdbcReservations {
+public class JdbcReservations implements Reservations {
 
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
